The start of something..
Alright so far we have a log which is modifed called application.log - going to be changed in the future to have date and time. just needed for quick testing.

Features

New

Open

Save

We intend in adding other file formats + syntax highlighting, formatting etc... Just as I have stated it is all a process.